Update paths_vs_bridges_methods.ipynb
I found an error in my covariance calculation that resolved the need for a correction to the covariance between sample nodes and nodes within the loop. Now functions as I would expect. Still need to fix the matrix in Nested Loops #2. Also, found a covariance matrix between sample nodes for the Split Bubble example that appears to match the Paths Method; worth more investigation.