Cleaned up Nested Bubbles #2 and #3. Progress on Split Bubble
Added ARG images for Nested Bubbles #2 and #3. Remade the covariance matrix for Nested Bubbles #2 because I initially miscalculated the covariances between sample nodes and internal nodes. Working on the Split Bubble ARG, I found covariance values that give a matching plot to the Paths Method, but I do not know why these values work. Hoping to come up with an equation similar to the Nested Bubbles that outputs these covariances.