Adding paths_modified and top_down_puneeth
Adding two new algorithms to the mix. paths_modified is a big speed improvement over the original paths method but keeps its simplicity. top_down_puneeth is still a work in progress as it doesn't yet handle "bubbles". Eventually, I would like to get rid of the the networkx dependency, which should be easier with top_down. The paths method is pretty reliant on nx.all_simple_paths().