Reworked the output of estimate_spatial_parameters()
`estimate_spatial_parameters()` didn't really flow well, especially if we wanted to calculate the location of an ancestor at an arbitrary position in the genome and time. I've separated `locate_ancestor()` from `estimate_spatial_parameters()`. The general workflow is that you estimate the spatial parameters and then use that output as an input to estimate the location of an ancestor. I've made this work with a pandas.DataFrame using `estimate_location()` which can be applied to each row of the DataFrame. This makes it easier to compare true values to estimated values. Variance in location estimates are also returned.