major restructuring and new figures
This will be the final time that I change the code in a significant way. I realized that all of the separated lists, dictionaries, and arrays that correspond to spatial parameters calculated for an ARG should really be brought into a single object. This does affect previous code so those will need to be rerun. I feel this is for the better and is much cleaner. It also means there is less that a user needs to do to replicate. Now the analysis is split into three steps: 1) cleaning and preparing the SLiM output for sparg, 2) initializing the sparg.SpatialARG and calculating all of the necessary spatial parameters, 3) estimating the locations of ancestors/comparing to other methods.