Improved windowing method and corrected variance
The windowing method now runs significantly faster as it first checks if there are overlapping windows that will use the same ARG. This avoid redundant computation of the covariance matrix. It does use more memory as it is storing a dictionary of the ARGs that are used more than once, but for our studies this does not seem to be a problem.

I've also added Puneeth's corrected variance calculation which handles multiroot ARGs (and trees). This includes a few more SpatialARG attributes which still need to be documented. There is still somewhat strange behavior, where the variance can decrease as we go back in the past, but this appears to be correct.